Remove some modules;
Vedeu::Repositories::Parent, Vedeu::Presentation::Colour::Background and
Vedeu::Presentation::Colour::Foreground have been subsumed into the classes
or module including them.
