+++ lib/vedeu/presentation/all.rb
@@ -11,8 +11,6 @@ module Vedeu
 
 end # Vedeu
 
-require 'vedeu/presentation/background'
-require 'vedeu/presentation/foreground'
 require 'vedeu/presentation/colour'
 require 'vedeu/presentation/styles'
 require 'vedeu/presentation/style'
