+++ lib/vedeu/repositories/all.rb
@@ -15,7 +15,6 @@ require 'vedeu/repositories/collection'
 require 'vedeu/repositories/defaults'
 require 'vedeu/repositories/repositories'
 require 'vedeu/repositories/model'
-require 'vedeu/repositories/parent'
 require 'vedeu/repositories/store'
 require 'vedeu/repositories/registerable'
 require 'vedeu/repositories/repository'
