+++ lib/vedeu/views/stream.rb
@@ -27,7 +27,6 @@ module Vedeu
       extend Forwardable
       include Vedeu::Views::DefaultAttributes
       include Vedeu::Repositories::Model
-      include Vedeu::Repositories::Parent
       include Vedeu::Presentation
 
       include Vedeu::Views::Value
@@ -87,6 +86,22 @@ module Vedeu
         self.class == other.class && value == other.value
       end
       alias_method :==, :eql?
+
+      # @return [NilClass|String|Symbol]
+      def name
+        if present?(@name)
+          @name
+
+        elsif parent && present?(parent.name)
+          parent.name
+
+        end
+      end
+
+      # @return [NilClass|void]
+      def parent
+        present?(@parent) ? @parent : nil
+      end
 
     end # Stream
